llm_AIO/requirements.txt
# llm_AIO 全项目依赖（网关、用户管理、多模态适配、数据集/转换、RAG）
# 安装: pip install -r requirements.txt

# ---------- Web 框架 ----------
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.5.0
starlette>=0.50.0
python-multipart>=0.0.21

# ---------- HTTP / 异步客户端 ----------
httpx>=0.24.0
requests>=2.31.0
aiohttp>=3.9.0

# ---------- 配置与校验 ----------
python-dotenv>=0.21.0
email-validator>=2.3.0

# ---------- 认证与安全 ----------
passlib>=1.7.4
python-jose[cryptography]>=3.5.0

# ---------- 系统监控 ----------
psutil>=7.2.1

# ---------- 数据库（含 LangGraph SQLite checkpoint 用的异步驱动）----------
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# ---------- 阿里云服务 ----------
dashscope>=1.20.0
oss2>=2.18.0

# ---------- RAG：LangGraph + Chroma + 检索工具 + 联网搜索 ----------
# LangChain 主包便于依赖解析；工具创建使用 langchain_core.tools.create_retriever_tool
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0,<4.0.0
chromadb>=0.5.0
pypdf>=4.0.0
ddgs>=6.0.0
# 与 OpenAI 兼容 API / LangChain 栈常用（显式列出，避免环境解析过旧）
openai>=1.40.0
tiktoken>=0.7.0
tenacity>=8.2.0
langsmith>=0.1.0
